@@ -1,56 +1,28 @@
 {
-  "name": "new_portfolio_typescript",
-  "version": "1.0.0",
+  "name": "new-portfolio",
   "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/cors": "^2.8.12",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^12.20.23",
-    "@types/nodemailer": "^6.4.4",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.1.8",
-    "@types/react-router-hash-link": "^2.4.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.1",
-    "react-router-hash-link": "^2.4.3",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.0",
-    "ts-node": "^10.2.1",
-    "typescript": "^4.3.5",
-    "web-vitals": "^1.1.2"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "lint": "eslint './src/**/*.{ts,tsx}'"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.8.0",
+    "react-router-hash-link": "^2.4.3",
+    "styled-components": "^5.3.6",
+    "vite-plugin-svgr": "^2.4.0"
   },
   "devDependencies": {
-    "@types/express": "^4.17.13",
-    "@types/styled-components": "^5.1.13"
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "@types/react-router-hash-link": "^2.4.5",
+    "@types/styled-components": "^5.1.26",
+    "@vitejs/plugin-react-swc": "^3.0.0",
+    "typescript": "^4.9.3",
+    "vite": "^4.1.0"
   }
 }
