@@ -1,16 +1,16 @@
-import React from "react";
-import { Switch, Route } from "react-router-dom";
-import HomePage from "./pages/Home";
-import Contact from "./pages/Contact";
-import Projects from "./pages/Projects";
+import { Routes, Route } from 'react-router-dom';
+
+import HomePage from './pages/Home';
+import Contact from './pages/Contact';
+import Projects from './pages/Projects';
 
 const PageRoutes = () => {
   return (
-    <Switch>
-      <Route exact path="/" component={HomePage} />
-      <Route path="/projects" component={Projects} />
-      <Route path="/contact" component={Contact} />
-    </Switch>
+    <Routes>
+      <Route path="/" element={<HomePage />} />
+      <Route path="/projects" element={<Projects />} />
+      <Route path="/contact" element={<Contact />} />
+    </Routes>
   );
 };
 
