@@ -1,18 +1,15 @@
-import React from "react";
-import { BrowserRouter as Router } from "react-router-dom";
-import { GlobalStyle, Theme } from "./globalandtheme";
-import PageRoutes from "./PageRoutes";
+import { BrowserRouter } from 'react-router-dom';
+import { GlobalStyle, Theme } from './globalandtheme';
+import PageRoutes from './PageRoutes';
 
 const App = () => {
   return (
-
-      <Router>
-        <GlobalStyle />
-        <Theme>
-          <PageRoutes />
-        </Theme>
-      </Router>
- 
+    <BrowserRouter>
+      <GlobalStyle />
+      <Theme>
+        <PageRoutes />
+      </Theme>
+    </BrowserRouter>
   );
 };
 
