@@ -1,10 +1,11 @@
-import styled from "styled-components";
-import background_image from "./assets/background-moon.png";
+import styled from 'styled-components';
+import background_image from './assets/background-moon.png';
 
 export const PancakeStack = styled.div`
   display: grid;
   grid-template-rows: 15vh auto;
   position: relative;
+  height: 100%;
 `;
 
 export const FadeInEffect = styled.div`
@@ -31,14 +32,17 @@ interface ICenterContainer {
 }
 
 export const CenterContainer = styled(FadeInEffect)<ICenterContainer>`
-  max-width: 35rem;
+  max-width: 45rem;
   text-align: center;
-  min-height: ${(props) => props.height};
-  @media only screen and (min-width: ${(props) => props.theme.mediaQuery.tabletWidth}) {
-    min-height: 130vh;
+  min-height: 100%;
+  div,
+  ul {
+    &:last-child {
+      padding-bottom: 45vh;
+    }
   }
   svg {
-    margin: 1rem;
+    margin: 0 1rem;
   }
 `;
 
@@ -47,9 +51,10 @@ export const MoonLayer = styled.div`
   overflow: hidden;
   background-size: cover;
   background-position: 100% 100%;
-  min-height: 55vh;
+  min-height: 100%;
   pointer-events: none;
   position: fixed;
+  z-index: 10;
   bottom: 0;
   left: 0;
   right: 0;
