@@ -1,17 +1,17 @@
-import React from "react";
-import { PancakeStack } from "../ComponentStyles";
-import Header from "../containers/Header";
-import MainContainer from "../containers/Main";
-import { MoonLayer } from "../ComponentStyles";
-import About from "../components/About";
+import React from 'react';
+import { PancakeStack } from '../ComponentStyles';
+import Header from '../containers/Header';
+import CenterSection from '../containers/CenterSection';
+import { MoonLayer } from '../ComponentStyles';
+import About from '../components/About';
 
 const Home = () => {
   return (
     <PancakeStack>
       <Header />
-      <MainContainer>
+      <CenterSection>
         <About />
-      </MainContainer>
+      </CenterSection>
       <MoonLayer />
     </PancakeStack>
   );
